requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
APScheduler==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
black==21.10b0
bs4==0.0.1
cachetools==4.2.2
calchundred @ file:///C:/Users/User/OneDrive/Desktop/pkg/dist/calchundred-0.2-py3-none-any.whl
calculator @ file:///C:/Users/User/OneDrive/Documents/PythonPackages/dist/calculator-0.3.0-py3-none-any.whl
certifi==2021.10.8
cffi==1.15.0
chainmap==1.0.3
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
combomethod==1.0.12
comtypes==1.1.10
discord==1.7.3
discord-py-slash-command==3.0.1
discord.py==1.7.3
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.0.2
future==0.18.2
idna==3.3
imageio-ffmpeg==0.4.5
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.2
Js2Py==0.71
MarkupSafe==2.0.1
MouseInfo==0.1.3
multidict==5.2.0
mypy-extensions==0.4.3
nextcord==2.0.0a3
nulltype==2.3.1
numpy==1.21.4
opencv-python==4.5.4.60
options==1.4.10
orjson==3.6.4
packagearinjoy @ file:///C:/Users/User/OneDrive/Desktop/packagearinjoy/dist/packagearinjoy-0.1-py3-none-any.whl
packagearinjoyn @ file:///C:/Users/User/OneDrive/Desktop/pkg/dist/packagearinjoyn-0.5-py3-none-any.whl
pandas==1.3.4
pathspec==0.9.0
pefile==2021.9.3
Pillow==8.4.0
platformdirs==2.4.0
plyer==2.0.0
py-cord @ git+https://github.com/Pycord-Development/pycord@e00b272d40b3462b12c6872ad0243a3fae2a43a2
PyAutoGUI==0.9.53
pycares==4.1.2
pycodestyle==2.8.0
pycord==0.1.1
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyperclip==1.8.2
pypiwin32==223
pypng==0.0.21
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.2
pywin32==302
pywin32-ctypes==0.2.0
pyzbar==0.1.8
regex==2021.11.10
requests==2.26.0
six==1.12.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
wikipedia==1.4.0
yarl==1.7.2
youtube-dl==2021.6.6
yt-search==1.1.0.post1
